Adding a distinct flag at fixtures to avoid duplicates (#42)
Solved: when the user plays La Liga (as League) and R.Madrid (As team league), we are returning Madrid's match twice.

For some reason, Ecto was filtering it... And it was not returning to the query. However, it was breaking the pagination, as the offset was greater than the returning list itself.

The distinct flag is ensuring we will no longer have this issue.